Fix problem with 'nil' attributes
Attributes that have a default value of nil where not enforcing its
default if they had been set to a different value in a previous run.

This is related to this issue: https://github.com/chef/chef/issues/1507
The workaround is to manually remove the attributes present in the node
after a cookbook run, so that in the next run the merge will always
override.